Fix doc links and add space (#417)
Previously the markdown links were merging the relative url with the url
of the current page, rather than the base url of the parent folder. E.g.
we were linking to architecture/architecture/consistent-hash-ring instead
of architecture/consistent-hash-ring).

Used relrefs for fixing links instead of modifying the markdown link, as
relrefs will throw errors if the page doesn't exist.