Fix object prefix not being stripped for Dell ECS
This commit ensure that the same object prefix string is used to query and strip
an object when listing all the objects in a bucket.

Some S3 implementations (in this case Dell ECS) cannot be relied upon to return
the correct `res.Prefix`. Dell ECS will return it URLencoded, e.g. "traceID%2F"
instead of "traceID/".