Filter out 'pause' container from operational dashboard
These metrics are typically very low because the "pause" container
is a piece of Kubernetes internals that sits and does nothing.
Filtering them out of the display aids readability.

Note in some Kubernetes installations the pause container is labled
"POD", which was previously filtered out from the dashboard.
See https://github.com/containerd/cri/issues/922 for background on the
name.