Don't set CompressResponses as this is a no-op
CompressResponses has been removed in v1.9.0 of Cortex. Additionally, this setting actually did not change anything when used in Tempo.

In Cortex this setting was used when creating the handler in query-frontend:
https://github.com/cortexproject/cortex/blob/4afaa357469fb37fd92cb1e2a0c1d10cdddc5ecd/pkg/cortex/modules.go#L557-L559
Since we do not use this code and have our own initialisation code for query-frontend, this setting never changed anything.