tempo-mixin: add panel for Envoy Proxy sidecar, update tanka example and jsonnet dependencies (#1137)
* tempo-mixin: add panel for Envoy Proxy sidecar, update tanka example and jsonnet dependencies

Changes:

### Tempo / Writes dashboard

Add a panel with Envoy metrics. This panel is to be used when the gateway is run with an Envoy Proxy sidecar (for HTTP2 load balancing).
The metrics from Envoy expose the gRPC status codes, which are essential pieces of information when debugging ingress issues. The gateway only reports HTTP statuses, which is deceptive since a 200 OK HTTP packet might contain a gRPC message with an error.
Example: if a request is denied by the ingester because it hits the limit, the distributor will return a HTTP 200 OK message containg a gRPC message with status 9 FAILED PRECONDITION.

I've also added a text panel listing the gRPC status codes since these aren't common knowledge.
It links to this doc: https://github.com/grpc/grpc/blob/master/doc/statuscodes.md

### tanka example

- update prometheus config to always add `cluster` label, the tempo-mixin dashboards depend on this label
- enable search in Grafana and Tempo
- use a relative link to tempo-mixin, this makes it easier to quickly test changes

I've also run `jb update` in tempo-mixin and tk examples, hence the large amount of vendor file being changed.

* Update CHANGELOG.md

* Regenerate tempo-mixin/yamls