Build and release Tempo for all GOOS/GOARCH combinations (#1510)
* Build tempo for all architectures
We are building a company-wide deb and RPM repository and we want to have all of our OSS products in it. Before I start working on this for Tempo, I saw that only linux-amd64 is built and released. This PR widens that, with the goal being of standardizing our offering across products (mimir, loki and grafana already offer binaries for all OSes/archs)

In this PR:

-Build all three tools (tempo-query, tempo-cli and tempo) across more archs and os combinations
- Keep the same archive format as now (all three tools in the same tar.gz file), let me know if that would be better to change. Though, changing the format could break automation that users have created
- Add a pull request run for make release-snapshot. Testing the whole releasing process. This will be very useful once I add rpm and deb packaging

* Put release snapshot in CI

* Release using Drone
The job takes much too long in Github Actions