Improve frontend and querier error handling
Here we aim to improve the error handling between the querier and the
frontend.  Several errors can be passed in either direction depending on
the situation, including context cancellation and gRPC "queue stopped"
messages.

In some cases, we simply avoid logging the error, recognizing it as
part of the normal connection lifecycle in worker-frontend relationship.
In the case a worker fails to contact the frontend, the worker will will
retry regardless of the error.  This behavior is unchanged.  Frontends
only get removed from the worker by the DNS record handling for the frontend address.

Additionally, a 1 second delay was added to the fronend shutdown.  This
is to allow the situation where both the frontend and querier modules
are running in the same binary.  Without this delay, there is a race
between the frontend and the querier to determine who can tell who a
shutdown has been requested.  This delay allows the query worker to
notify the frontend of a shutdown request so that both components can
shut down cleanly.