Refactor poller into new module and adjust tempodb.New everywhere
The poller has write operations which, if not canceled, have unknown
behavior.  In order to pass the service context into the poller, this
change proposes moving the poller itself into a module that can be used
as a dependency and allow for more control over the termination of the
write operations to the backend.

Also included here is a rename of the instance variables as well as the
instance type of tempoDB.  I believe this is more clear as to the
intention.

Structurally, the `tempodb.New` now returns the `tempoDB` instance.
This allows the caller to use the returned instance as whatever
interface is desired. Adjustments for this contract have been updated
everywhere.

Additional formatting changes for `gofumpt` taking over my editor.