microservices: use DNS lookup for memberlsit.join_members by default
Related: https://github.com/grafana/loki/pull/9723

In high instance environments, the current default configuration will result
in a connection to each member returned by a DNS entry and result in
more connections than is necessary.

Here we include the fix that Loki has for using a `dns+` lookup feature
of memberlist to ensure that only a few members of the DNS entry are
added.