Fix tools package listing
Without this change, the `go list` used to track with tools are required
by the project is broken in Go 1.21, resulting in the following error.

```
tools.go:6:2: import "github.com/psampaz/go-mod-outdated" is a program, not an importable package
```

Here we add `-e` to the `go list` to allow the list to complete and
handle the tool management as before.