Set GRAFANA_VERSION used throughout shared lookups (#3091)
It was working in "next" because the fallback behavior is to use the page's version for any "SOMETHING_VERSION" lookup.
Therefore, the `docs/shared` lookup coincidentally targeted "next" which exists in the Grafana documentation.

Now that we have cut a new release with these docs, the released version is being used which is causing errors because `v2.3.x` is not a valid Grafana version.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>
(cherry picked from commit 32014d78675fba33f47b2f9627c4df53be0a35d5)