Refactor InspectedBytes collection
Refactor metrics collection for InspectedBytes to ensure that we do it in the same way in all endpoints.
This also adds the metrics collection in some places where it was missing, which is probably why sometimes
we get 0 reported bytes.
In addition, avoid usage of an atomic where not necessary for performance.