Add jitter in backendworker to avoid thundering herd from workers (#5150)
Right now we can have a case where the scheduler comes up, all the workers connect to it and wait for jobs and fails due to issues or when the scheduler has no jobs or failed to handout jobs to workers. all the workers will start backing off with retry with the same backoff durations.

This can cause waves of thundering herd on the scheduler if we have a sizeable amount of workers.

This PR add a jitter in workers so they are offset by jitter and won't cause waves of thundering herd on the scheduler.

This is even more important because backend scheduler is a singleton component and can be easily overwhelmed in bigger clusters.

Jitter code in play: go.dev/play/p/qb5GU11lKF5