linter: disable gosec rule g115
G115: Potential integer overflow when converting between integer types

This rule introduced a lot of noise as we have a lot of cases
where we truncate or change the sign of an integer type. In addition,
in many cases protobuf definitions are involved, which require regeneration.
But most importantly, `len` always returns `int`, which means we should
add a conditional every time we need to convert it, potentially affecting performance.