[backend-scheduler] fix duplicate work handout during race
A race exists between the compaction provider and the Next call of the
scheduler where the provider may detect the same block IDs for a job for
which the cache has not yet been updated.  This results in the failed
jobs and duplicate data in the backend due to two compactions happening
on the same blocks.