[Generator] Add entity limiter
This PR adds a new type of limiter to the metrics generator. It is
designed to limit on "entities" rather than series. In this case an
entity is a single labelset across multiple metrics, excluding external
labels. In effect, it allows the limiter to always generate the full set
of data for a given entity, rather than limiting randomly once the
series limit is triggered.

Importantly, this limit is applied at collection time, so it's still
necessary to set a series limit in order to avoid generator OOM kills.
In practice, the series limit should likely be some constant multiple of
the entity limit. e.g. 20x. This allows the system to primarily limit on
entities while still enforcing a ceiling on memory consumption.