docs: why we use gogo protobuf
The `gogo/protobuf` library is deprecated and abandoned; it has not
received any updates in over 4.5 years. As such, it may appear peculiar
that we are still dependent on it, rather than some other solution.

At some point, someone else will surely also ask this same question,
just like I did. This implements a retroactive design proposal to
document why we use this fork, and what we require in a replacement.
(These requirements are not currently fulfilled by a single alternative.)