Fix delete implementation for s3/gcs/azure backends to account for prefix (#6011)
* Fix delete impl for s3/gcs/azure backends to account for prefix

* Update changelog

* Omit prefix from backend.readerWriter.Find result set

The backend delete function (which prepends the prefix) is being used
in two places: first to delete indexes directly using the file path,
and secondly, the results from the Find function are iterated over to
delete the files for a tenant. If the result set from the Find
function returns the full path (including the prefix), the delete
function would prepend the prefix again and hence would not target the
correct path. Hence all Find implementations are updated to return
unprefixed key in the result set (we trim the prefix from key before passing
over to the FindFunc arg). This is consistent with the List implementation.

* Add prefix for keypath for versioned deletes

* Remove bucket prefix in backend.Find instead of entire prefix